use get_plugins to decide whether or not to add oafeat-link
